util/ioprocs.cpp: Added wrappers for common patterns. (#11608)
emu/diimage.h: Removed fread overloads that allocate memory for output.

util/core_file.cpp: Changed output size of load to size_t.

+++ src/lib/util/corefile.h
@@ -77,7 +77,7 @@ public:
 	virtual char *gets(char *s, int n) = 0;
 
 	// open a file with the specified filename, read it into memory, and return a pointer
-	static std::error_condition load(std::string_view filename, void **data, std::uint32_t &length) noexcept;
+	static std::error_condition load(std::string_view filename, void **data, std::size_t &length) noexcept;
 	static std::error_condition load(std::string_view filename, std::vector<uint8_t> &data) noexcept;
